+++ front/cypress/e2e/survey_builder/utils.ts
@@ -1,0 +1,45 @@
+import { randomString } from '../../support/commands';
+import moment = require('moment');
+
+export const createSurveyProject = (cy: any) => {
+  const projectTitle = randomString();
+  const phaseTitle = randomString();
+  const projectDescription = randomString();
+  const projectDescriptionPreview = randomString(30);
+  let projectId: string;
+  let projectSlug: string;
+  let phaseId: string;
+
+  return new Promise((res) => {
+    cy.setAdminLoginCookie();
+    cy.apiCreateProject({
+      title: projectTitle,
+      descriptionPreview: projectDescriptionPreview,
+      description: projectDescription,
+      publicationStatus: 'published',
+    })
+      .then((project: any) => {
+        projectId = project.body.data.id;
+        projectSlug = project.body.data.attributes.slug;
+        return cy.apiCreatePhase({
+          projectId: projectId as string,
+          title: phaseTitle,
+          startAt: moment().subtract(9, 'month').format('DD/MM/YYYY'),
+          participationMethod: 'native_survey',
+          nativeSurveyButtonMultiloc: { en: 'Take the survey' },
+          nativeSurveyTitleMultiloc: { en: 'Survey' },
+          canPost: true,
+          canComment: true,
+          canReact: true,
+        });
+      })
+      .then((phase: any) => {
+        phaseId = phase.body.data.id;
+        res({
+          projectId,
+          projectSlug,
+          phaseId,
+        });
+      });
+  });
+};
